Created content for original runewords.
This took a bit of thinking. The runewords are contain within them combinations
of every other model. The problem here, is that properties are not listed anywhere
for me to easily recreate them. I had to create the properties while populating the
runewords. It took a bit of organizational setup, but eventually I got in the
groove and started hammering away.

I chose yml format because it was the more compact and sensible than json. We do want
to expose json of course, but to import the data (which is static) I chose something
simple and which I am familiar with.

Next step, rake task to model create